requirements.txt
# Core
streamlit==1.35.0
python-dotenv==1.0.1

# Data Stack
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.4
scikit-learn==1.5.0

# Visualization
plotly==5.22.0
seaborn==0.13.2
matplotlib==3.8.4

# LangChain Ecosystem
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.2

# Vector DB
faiss-cpu==1.8.0

# LLM Providers
openai==1.35.3
google-generativeai>=0.5.2,<0.6.0
groq==0.9.0
tiktoken==0.7.0

# File Handling
pypdf==4.2.0
pdfplumber==0.11.1

# Auth & DB
bcrypt==4.1.3
sqlalchemy==2.0.30

# Embeddings
sentence-transformers==3.0.1